Get BIP32 & DNS unit tests to run, and fix compilation errors.

+++ cppForSwig/BinaryData.h
@@ -253,11 +253,11 @@ public:
 
 
    /////////////////////////////////////////////////////////////////////////////
-   BinaryData & append(uint8_t const * str, uint32_t sz);
+   BinaryData & append(uint8_t const * str, size_t sz);
 
 
    /////////////////////////////////////////////////////////////////////////////
-   BinaryData & append(uint8_t const byte, uint32_t sz)
+   BinaryData & append(uint8_t const byte, size_t sz)
    {
       for(uint32_t x = 0; x < sz; ++x)
       {
